Execution & Memory Spaces

Kokkos.Spaces.Space — Type

Space

Abstract super type of all execution and memory spaces.

Main subtypes:

ExecutionSpaces: super type of all execution spaces

MemorySpaces: super type of all memory spaces

All Kokkos spaces have a main abstract type (Serial, Cuda, HostSpace, HIPSpace...) which are defined even if it has not been compiled on the C++ side. Those main abstract types should be the ones used when specifying a space. This allows methods like enabled to work independently from the compiled internal library.

When a space is compiled, a sub-type of its main type is defined by CxxWrap, leading to the following type structure: SerialImplAllocated <: SerialImpl <: Serial <: ExecutionSpace <: Space. Below the main space type (here, Serial), the sub-types are only defined if they are compiled, and therefore they should not be relied upon.

Navigating the type tree can be made easier through main_space_type.

Kokkos.Spaces.ExecutionSpace — Type

ExecutionSpace <: Space

Abstract super-type of all execution spaces.

Sub-types:

Serial

OpenMP

OpenACC

OpenMPTarget

Threads

Cuda

HIP

HPX

SYCL

All sub-types are always defined, but only some of them are enabled. To enable an execution space, you must enable its related Kokkos backend, e.g. '-DKokkosENABLESERIAL=ON' for the Serial execution space.

To do this you can set the backends option with Kokkos.set_backends, or specify the option directly through kokkos_options.

Kokkos.Spaces.MemorySpace — Type

MemorySpace <: Space

Abstract super-type of all memory spaces.

Sub-types:

HostSpace

CudaSpace

CudaHostPinnedSpace

CudaUVMSpace

HIPSpace

HIPHostPinnedSpace

HIPManagedSpace

Sub-types work the same as for ExecutionSpace. They can be enabled by enabling their respective backend.

Kokkos.Spaces.accessible — Function

accessible(::View)

Return true if the view is accesssible from the default host execution space.

accessible(::Union{<:MemorySpace, Type{<:MemorySpace}})

Return true if the memory space is accessible from the default host execution space.

Equivalent to Kokkos::SpaceAccessibility<Kokkos::DefaultHostExecutionSpace, Space>::accessible

Kokkos.Spaces.enabled — Function

enabled(space::Union{Space, Type{<:Space}})

Return true if the given execution or memory space is compiled.

Kokkos.Spaces.execution_space — Function

execution_space(space::Union{<:MemorySpace, Type{<:MemorySpace}})

Return the execution space associated by default to the given memory space.

Kokkos.Spaces.memory_space — Function

memory_space(space::Union{<:ExecutionSpace, Type{<:ExecutionSpace}})

Return the memory space associated by default to the given execution space.

Kokkos.Spaces.main_space_type — Function

main_space_type(space::Union{<:Space, Type{<:Space}})

Return the main space type of space, e.g. for Serial, SerialImpl or SerialImplAllocated we get Serial.

Kokkos.Spaces.kokkos_name — Function

kokkos_name(space::Union{Space, Type{<:Space}})

Return the name of the execution or memory space as defined by Kokkos.

Equivalent to Kokkos::space::name()

Constants

Kokkos.Spaces.COMPILED_EXEC_SPACES — Constant

COMPILED_EXEC_SPACES::Tuple{Vararg{Type{<:ExecutionSpace}}}

List of all compiled Kokkos execution spaces.

Kokkos.Spaces.COMPILED_MEM_SPACES — Constant

COMPILED_MEM_SPACES::Tuple{Vararg{Type{<:MemorySpace}}}

List of all compiled Kokkos execution spaces.

Default spaces

Kokkos.Spaces.DEFAULT_DEVICE_SPACE — Type

DEFAULT_DEVICE_SPACE::Type{<:ExecutionSpace}

The default execution space where kernels are applied on the device.

Equivalent to Kokkos::DefaultExecutionSpace.

Kokkos.Spaces.DEFAULT_DEVICE_MEM_SPACE — Type

DEFAULT_DEVICE_MEM_SPACE::Type{<:MemorySpace}

The default memory space where views are stored on the device.

Equivalent to Kokkos::DefaultExecutionSpace::memory_space.

Kokkos.Spaces.DEFAULT_HOST_SPACE — Type

DEFAULT_HOST_SPACE::Type{<:ExecutionSpace}

The default execution space where kernels are applied on the host.

Equivalent to Kokkos::DefaultHostExecutionSpace.

Kokkos.Spaces.DEFAULT_HOST_MEM_SPACE — Type

DEFAULT_HOST_MEM_SPACE::Type{<:MemorySpace}

The default memory space where views are stored on the host.

Equivalent to Kokkos::DefaultHostExecutionSpace::memory_space.

Kokkos.Spaces.SHARED_MEMORY_SPACE — Type

SHARED_MEMORY_SPACE::Union{Nothing, Type{<:MemorySpace}}

The shared memory space between the host and device, or nothing if there is none.

Equivalent to Kokkos::SharedSpace if Kokkos::has_shared_space is true.

Kokkos.Spaces.SHARED_HOST_PINNED_MEMORY_SPACE — Type

SHARED_HOST_PINNED_MEMORY_SPACE::Union{Nothing, Type{<:MemorySpace}}

The shared pinned memory space between the host and device, or nothing if there is none.

Equivalent to Kokkos::SharedHostPinnedSpace if Kokkos::has_shared_host_pinned_space is true.