Views

View

Kokkos.Views.View — Type

View{T, D, MemSpace} <: AbstractArray{T, D}

Wrapper around a Kokkos::View of D dimensions of type T, stored in MemSpace.

Behaves like a normal Array. Indexing is done by calling the Kokkos::View::operator() function of the view, and is therefore not highly performant. The best performance with Kokkos views is achieved by calling Kokkos kernels compiled from C++.

It is supposed that all view accesses are done from the default host execution space. Since the view may be stored in a memory space different from the host, it may be invalid to access its elements: if accessible(MemSpace) is false, then all view accesses will throw an error.

Views are created through CxxWrap.jl, which adds automatically a finalizer to all objects which calls the view's destructor when the Julia object is deleted by the garbage collector.

It is important to understand that for a view to be properly disposed of, there is two requirements:

the library containing its destructor is still loaded. Views created by the Kokkos.jl library will always meet this requirement.

finalize wasn't called.

Kokkos.Spaces.accessible — Method

accessible(::View)

Return true if the view is accesssible from the default host execution space.

Kokkos.Views.label — Function

label(::View)

Return the label of the View.

Kokkos.Views.view_data — Function

view_data(::View)

The pointer to the data of the View.

Kokkos.Views.view_wrap — Function

view_wrap(array::DenseArray{T, D})
view_wrap(::Type{View{T, D, S}}, array::DenseArray{T, D})
view_wrap(::Type{View{T, D, S}}, d::NTuple{D, Int}, p::Ptr{T})

Construct a new View from the data of a Julia-allocated array. The returned view does not own the data, and no copy is made.

The memory space S defaults to DEFAULT_HOST_MEM_SPACE.

Warning

Julia arrays have a column-major layout. Kokkos can handle both row and column major array layouts, but it is imposed by the memory space. If the layouts don't match the view behaviour is unpredictable. This only affects 2D arrays and above.

Warning

The returned view does not hold a reference to the original Julia array. It is the responsability of the user to make sure the original array is kept alive as long as the view should be accessed.

Constants

Kokkos.Views.Idx — Type

Idx::Type{<:Integer}

Integer type used by views for indexing on the default execution device. Usually either Cint or Clonglong.

Equivalent to Kokkos::RangePolicy<>::index_type.

Kokkos.Views.COMPILED_DIMS — Constant

COMPILED_DIMS::Tuple{Vararg{Int32}}

List of all View dimensions which are compiled.

By default, dimensions 1 and 2 are compiled.

Kokkos.Views.COMPILED_TYPES — Constant

COMPILED_TYPES::Tuple{Vararg{DataType}}

List of all View element types which are compiled

By default, the following types are compiled: Float64 (double), Float32 (float) and Int64 (int64_t).