Environment

Initialization

Kokkos.initialize — Function

initialize(;
    num_threads=nothing,
    device_id=nothing, map_device_id_by=nothing,
    disable_warnings=nothing, print_configuration=nothing,
    tune_internals=nothing,
    tools_libs=nothing, tools_args=nothing
)

Initializes Kokkos by calling Kokkos::initialize().

The keyword arguments build are passed to the InitializationSettings constructor which is then passed to Kokkos::initialize(). A value of nothing keeps the default behaviour of Kokkos.

The Kokkos wrapper library is loaded (and recompiled if needed) if it is not already the case. This locks the current Configuration Options until the end of the current Julia session.

Kokkos.is_initialized — Function

is_initialized()

Return Kokkos::is_initialized().

Can be called before the wrapper library is loaded.

Kokkos.finalize — Function

finalize()

Calls Kokkos::finalize().

Warning

Kokkos requires that all view destructors should be called before calling finalize.

Kokkos.is_finalized — Function

is_finalized()

Return Kokkos::is_finalized().

Can be called before the wrapper library is loaded.

Kokkos.KokkosWrapper.load_wrapper_lib — Function

load_wrapper_lib(; no_compilation=false, no_git=false, loading_bar=true)

Configures, compiles then loads the wrapper library using the current Configuration Options.

After calling this method, all configuration options become locked.

If no_compilation is true, then the CMake project of the wrapper library will not be configured or compiled.

If no_git is true, then if we need to use the Kokkos installation of the package, no Git operations (clone + checkout) will be done.

Both no_compilation=true and no_git=true are needed when initializing Kokkos.jl in non-root MPI processes.

Kokkos.is_kokkos_wrapper_loaded — Function

is_kokkos_wrapper_loaded()

Return true if load_wrapper_lib has been called.

Configuration

Kokkos.build_in_scratch — Function

build_in_scratch()

Sets the build_dir configuration option to a scratch directory (the default).

Kokkos.build_in_tmp — Function

build_in_tmp()

Sets the build_dir configuration option to a temporary directory, cleaned at the end of the Julia session.

Kokkos.build_in_project — Function

build_in_project(name = ".kokkos-build")

Sets the build_dir configuration option to the directory name in the active project directory.

Kokkos.require — Function

require(;
    version=nothing,
    dims=nothing, types=nothing, layouts=nothing,
    idx=nothing,
    exec_spaces=nothing, mem_spaces=nothing,
    no_error=false
)

Asserts that the underlying Kokkos wrapper library of Kokkos.jl respects the given requirements.

An argument with a value of nothing is considered to have no requirements.

version checks for the version of Kokkos.

idx checks the type of the index variables used when accessing a view.

version and idx are given as callables returning a Bool and taking a single argument: respectively a VersionNumber and a Type, e.g. passing version = >=(v"4.0.0") will match all Kokkos versions including v4.0.0 and above.

dims, types, layouts, exec_spaces and mem_spaces are lists of the required values.

dims, types and layouts check the available dimensions, types and layouts of views, while exec_spaces and mem_spaces do the same for execution and memory spaces.

If no_error is true, then this function will return false if any requirement is not met.

This function does not require for Kokkos to be initialized, but for the wrapper library to be loaded. If the wrapper is not loaded, the configuration options will be checked instead, however they cannot cover all possible requirements (idx and mem_spaces do not work and version works only if the packaged Kokkos installation is used).

Examples

# Require Kokkos version 4.0.0 (exactly), and for 1D and 2D views of Float64 to be compiled with
# a column or row major layout:
Kokkos.require(;
    version = ==(v"4.0.0"),
    types = [Float64],
    dims = [1, 2],
    layouts = [Kokkos.LayoutLeft, Kokkos.LayoutRight]    
)

# Require an index type of 8 bytes, the Cuda and OpenMP backends of Kokkos, as well as the Cuda UVM
# memory space to be available:
Kokkos.require(;
    idx = (==(8) ∘ sizeof),
    exec_spaces = [Kokkos.Cuda, Kokkos.OpenMP],
    mem_spaces = [Kokkos.CudaUVMSpace]
)

Kokkos.configinfo — Function

configinfo(io::IO = stdout)

Print information about all configuration options.

Kokkos.versioninfo — Function

versioninfo(io::IO = stdout; internal=true, verbose=false)

Print the version and various information about the underlying Kokkos library.

If internal is true, Kokkos::print_configuration() is called. verbose is passed to that call.

This function does not require for Kokkos to be initialized, however if internal=true then the output will be incomplete.

Kokkos.print_configuration — Function

print_configuration(io::IO, verbose::Bool)

Prints the internal Kokkos configuration to io.

Equivalent to Kokkos::print_configuration(out, verbose).

Kokkos.set_omp_vars — Function

set_omp_vars(;
    places = "cores",
    bind = "close",
    num_threads = Threads.nthreads()
)

Helper function to set the main OpenMP environment variables used by Kokkos. It must be called before calling initialize.

places sets OMP_PLACES. bind sets OMP_PROC_BIND. num_threads sets OMP_NUM_THREADS.

Note that since Julia threads and OpenMP threads are decoupled, there is no constraint imposed by Julia on OpenMP threads: there can be as many threads as needed.

Warning

Pinning the Julia threads with ThreadPinning.jl or with the JULIA_EXCLUSIVE environment variable can have an impact on OpenMP thread affinities, making the OpenMP variables useless.

Synchronization

Kokkos.fence — Method

fence()
fence(label::String)

Wait for all asynchronous Kokkos operations to complete.

Equivalent to Kokkos::fence().

Constants

Kokkos.KOKKOS_VERSION — Constant

KOKKOS_VERSION::VersionNumber

The Kokkos version currently loaded.

nothing if Kokkos is not yet loaded. See kokkos_version for the version of the packaged installation of Kokkos, which is defined before loading Kokkos.