Views

View

Kokkos.Views.View — Type

View{T, D, Layout, MemSpace} <: AbstractArray{T, D}

Wrapper around a Kokkos::View of D dimensions of type T, stored in MemSpace using the Layout.

Behaves like a normal Array. Indexing is done by calling the Kokkos::View::operator() function of the view, and is therefore not highly performant. The best performance with Kokkos views is achieved by calling Kokkos kernels compiled from C++.

It is supposed that all view accesses are done from the default host execution space. Since the view may be stored in a memory space different from the host, it may be invalid to access its elements: if accessible(MemSpace) is false, then all view accesses will throw an error.

Views are created through CxxWrap.jl, which adds automatically a finalizer to all objects which calls the view's destructor when the Julia object is deleted by the garbage collector.

It is important to understand that for a view to be properly disposed of, there is two requirements:

the library containing its destructor is still loaded. Views created by the Kokkos.jl library will always meet this requirement.

finalize wasn't called.

Kokkos.Spaces.accessible — Method

accessible(::View)

Return true if the view is accessible from the default host execution space.

Kokkos.Spaces.memory_space — Method

memory_space(::View)

The memory space type in which the view data is stored.

julia> my_cuda_space = Kokkos.CudaSpace()
 ...

julia> v = View{Float64}(undef, 10; mem_space=my_cuda_space)
 ...

julia> memory_space(v)
Kokkos.Spaces.CudaSpace

Kokkos.Spaces.array_layout — Method

array_layout(::View)

Return the Layout type of the view.

Kokkos.Views.label — Function

label(::View)

Return the label of the View.

Kokkos.Views.view_data — Function

view_data(::View)

The pointer to the data of the View. Using Base.pointer(view) is preferred over this method.

Equivalent to view.data().

Kokkos.Views.memory_span — Function

memory_span(::View)

Total size of the view data in memory, in bytes.

Equivalent to view.impl_map().memory_span().

Kokkos.Views.view_wrap — Function

view_wrap(array::DenseArray{T, D})
view_wrap(::Type{View{T, D, L, S}}, array::DenseArray{T, D})
view_wrap(::Type{View{T, D, L, S}}, d::NTuple{D, Int}, p::Ptr{T}; layout = nothing)

Construct a new View from the data of a Julia-allocated array. The returned view does not own the data, and no copy is made.

The memory space S defaults to DEFAULT_HOST_MEM_SPACE, and the layout L to array_layout(S).

If L is LayoutStride, then the kwarg layout should be an instance of a LayoutStride which specifies the stride of each dimension.

Warning

Julia arrays have a column-major layout by default. This correspond to a LayoutLeft, while Kokkos prefers LayoutRight for CPU allocated arrays. If strides(array) ≠ strides(view_wrap(array)) then it might lead to segfaults. This only concerns 2D arrays and above.

Warning

The returned view does not hold a reference to the original Julia array. It is the responsability of the user to make sure the original array is kept alive as long as the view should be accessed.

Kokkos.Views.deep_copy — Function

deep_copy(dest::View, src::View)
deep_copy(space::ExecutionSpace, dest::View, src::View)

Performs a copy of all data from src to dest.

In order for the copy to be possible, both views must have the same dimension, and either have the same layout or are both accessible from space.

If a space is given, the copy may be asynchronous. If not the copy will be synchronous.

Equivalent to Kokkos::deep_copy(dest, src) or Kokkos::deep_copy(space, dest, src). See the Kokkos docs about Kokkos::deep_copy

Kokkos.Views.create_mirror — Function

create_mirror(src::View; mem_space = nothing, zero_fill = false)

Create a new View in the same way as similar(src), with the same layout and padding as src.

If mem_space is nothing the new view will be in a memory space accessible by the host, otherwise it must be a memory space instance where the new view will be allocated.

If zero_fill is true, the new view will have all of its elements set to their default value.

See the Kokkos docs about Kokkos::create_mirror

Kokkos.Views.create_mirror_view — Function

create_mirror_view(src::View; mem_space = nothing, zero_fill = false)

Equivalent to create_mirror, but if src is already accessible by the host, src is returned and no view is created.

Kokkos.Views.impl_view_type — Function

impl_view_type(::Type{View{T, D, L, S}})

Returns the internal View type for the given complete View type.

julia> view_t = Kokkos.View{Float64, 2, Kokkos.LayoutRight, Kokkos.HostSpace}
Kokkos.Views.View{Float64, 2, Kokkos.Views.LayoutRight, Kokkos.Spaces.HostSpace}

julia> view_impl_t = Kokkos.impl_view_type(view_t)
Kokkos.KokkosWrapper.Impl.View2D_R_HostAllocated{Float64}

julia> supertype(supertype(view_impl_t))
Kokkos.Views.View{Float64, 2, Kokkos.Views.LayoutRight, Kokkos.KokkosWrapper.Impl.HostSpaceImplAllocated}

julia> view_impl_t <: view_t  # Julia types are contra-variant
false

Layouts

Kokkos.Views.Layout — Type

Layout

Abstract super-type of all view layouts.

Sub-types:

LayoutLeft

LayoutRight

LayoutStride

Kokkos.Views.LayoutLeft — Type

LayoutLeft

Fortran-style column major array layout. This is also the layout of Julia arrays.

While indexing, the first index is the contiguous one: v[i0, i1, i2].

Equivalent to Kokkos::LayoutLeft.

Kokkos.Views.LayoutRight — Type

LayoutRight

C-style row major array layout.

While indexing, the last index is the contiguous one: v[i2, i1, i0].

Equivalent to Kokkos::LayoutRight.

Kokkos.Views.LayoutStride — Type

LayoutStride

Arbitrary array layout, mostly used for sub-views.

Equivalent to Kokkos::LayoutStride.

When building a new view with a LayoutStride, the strides of each dimension must be given to the view constructor:

# A 7×11 matrix with column-major layout
v = Kokkos.View{Float64}(undef, 7, 11; layout=LayoutStride(1, 11))

# A 7×11 matrix with row-major layout
v = Kokkos.View{Float64}(undef, 7, 11; layout=LayoutStride(7, 1))

This differs slightly from the C++ Kokkos constructor, where dimensions and strides are interleaved.

Constants

Kokkos.Views.Idx — Constant

Idx::Type{<:Integer}

Integer type used by views for indexing on the default execution device. Usually either Cint or Clonglong.

Equivalent to Kokkos::RangePolicy<>::index_type.

nothing if Kokkos is not yet loaded.

Kokkos.Views.COMPILED_DIMS — Constant

COMPILED_DIMS::Tuple{Vararg{Int32}}

List of all View dimensions which are compiled.

By default, dimensions 1 and 2 are compiled.

nothing if Kokkos is not yet loaded.

Kokkos.Views.COMPILED_TYPES — Constant

COMPILED_TYPES::Tuple{Vararg{DataType}}

List of all View element types which are compiled

By default, the following types are compiled: Float64 (double), and Int64 (int64_t).

nothing if Kokkos is not yet loaded.

Kokkos.Views.COMPILED_LAYOUTS — Constant

COMPILED_LAYOUTS::Tuple{Vararg{DataType}}

List of all Layouts types which are compiled.

By default, the default array layout for the device and host execution spaces are compiled.

nothing if Kokkos is not yet loaded.